data-predictor/app/requirements.txt
influxdb
pyyaml